[demo] Callbacks really aren't working out
I'm not sure what is going on with SWIG, but everything I'm trying with
the callbacks results in a crash inside them as soon as one tries to
call a python function from within the callback.  That means that things
like:

  python -> c++

are ok, whereas:

  python -> c++ -> python

crashes, as does:

  python -> c++ -> python -> c++

(for subtly different reasons). For now, it's easier just to poll the
GameModelManager for the needed data. I conducted a quick experiment to
see what the poll sampling rate was, and it appeared to be around 1800
samples / second... which is plenty of headroom for other computation
(i.e. sending data to the ATEM).